test(symbolicator): Apply fix for location flakey tests.\n\nFixes #44559. Same approach as #44573.
