+++ tests/symbolicator/__init__.py
@@ -95,3 +95,19 @@ def insta_snapshot_stacktrace_data(self, event, **kwargs):
         },
         **kwargs,
     )
+
+
+def redact_location(candidates):
+    """Redacts the sentry location URI to be independent of the specific ID.
+
+    This modifies the data passed in, returns None.
+    """
+    location_re = re.compile("^sentry://project_debug_file/[0-9]+$")
+    for candidate in candidates:
+        try:
+            location = candidate["location"]
+        except KeyError:
+            continue
+        else:
+            if location_re.search(location):
+                candidate["location"] = "sentry://project_debug_file/x"
